Create README.md
As of submission time, I have not been able to find a solution to my if else statements. As I have commented in my .js file, an answer to the second question reverts to the first question, no matter which linked questions I use (ie: question 1 and 2, 2 and 4, 1 and 3, etc). 

I know there is something I'm not understanding about the if else statement, like I'm missing a way to make the process loop back around to the beginning (if that's even possible). Logically, I believe I have these in the proper order, but there is still so much about Javascript I don't understand. I'm going to keep investigating through the long weekend and will resubmit next week.